app/requirements.txt
blinker==1.9.0
click==8.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
opencv-python==4.13.0.92
pillow==12.1.1
psycopg2-binary==2.9.11
Werkzeug==3.1.6
